removed automatic import of handler object in events package and renamed handler.py to event_handler.py to solve name collision
