+++ kube_hunter/core/events/__init__.py
@@ -1,3 +1,2 @@
 # flake8: noqa: E402
-from .handler import EventQueue, handler
 from . import types
