+++ kube_hunter/core/types/hunters.py
@@ -19,7 +19,7 @@ class HunterBase:
     def publish_event(self, event):
         # Import here to avoid circular import from events package.
         # imports are cached in python so this should not affect runtime
-        from ..events import handler  # noqa
+        from ..events.event_handler import handler  # noqa
 
         handler.publish_event(event, caller=self)
